#feature Finished deleteDeadline method

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/service/DeadlineService.java
@@ -2,10 +2,13 @@ package nz.ac.canterbury.seng302.portfolio.service;
 
 import nz.ac.canterbury.seng302.portfolio.model.Deadline;
 import nz.ac.canterbury.seng302.portfolio.model.DeadlineRepository;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.PersistenceException;
 import java.util.Optional;
+import org.slf4j.Logger;
 
 /**
  * Client service used to communicate to the database and perform business logic for Deadlines
@@ -13,6 +16,7 @@ import java.util.Optional;
 @Service
 public class DeadlineService {
     @Autowired private DeadlineRepository deadlineRepository;
+    private Logger logger = LoggerFactory.getLogger(DeadlineService.class);
 
     /**
      * Creates a new deadline
@@ -36,8 +40,25 @@ public class DeadlineService {
             throw new IncorrectDetailsException("Failed to locate deadline in the database");
     }
 
-    public void deleteDeadline(int deadlineId) {
-        deadlineRepository.deleteById(deadlineId);
+    /**
+     * Delets deadline object from the database
+     * @param deadlineId
+     */
+    public String deleteDeadline(int deadlineId) throws IncorrectDetailsException, PersistenceException {
+        try {
+            Optional<Deadline> deadline = deadlineRepository.findById(deadlineId);
+            if(deadline.isPresent()) {
+                deadlineRepository.deleteById(deadlineId);
+                return "Successfully deleted " + deadline.get().getName();
+            }
+            else {
+                throw new IncorrectDetailsException("Could not find given Deadline");
+            }
+
+        } catch (PersistenceException e) {
+            logger.error("Failure deleting Deadline", e);
+            throw new IncorrectDetailsException("Failure deleting Deadline");
+        }
     }
 
 }
